Move profile module to top level
Placing this in the diagnostics submodule introduced dependency cycles
